+++ packages/waveform-desktop/scripts/install-electron.cjs
@@ -1,0 +1,5 @@
+'use strict';
+
+if (process.env.VERIFLOW_SKIP_ELECTRON_DOWNLOAD !== '1') {
+    require('electron');
+}